claudecode/architecture_intelligence/requirements.txt
# Architecture Intelligence Platform Dependencies

# Core dependencies
click>=8.0.0                    # CLI framework
rich>=13.0.0                   # Rich terminal output
pyyaml>=6.0                    # YAML configuration support
pydantic>=2.0.0                # Data validation and settings
asyncio>=3.4.3                 # Asynchronous programming

# Intelligence and ML
chromadb>=0.4.0                # Vector database for knowledge storage
sentence-transformers>=2.2.0   # Semantic similarity
numpy>=1.21.0                  # Numerical computing
pandas>=1.3.0                  # Data manipulation

# Framework-specific dependencies
plantuml-markdown>=3.5.0       # UML diagram generation
graphviz>=0.20                 # Graph visualization
mermaid-py>=1.0.0              # Mermaid diagram support

# Architecture documentation
mkdocs>=1.4.0                  # Documentation generation
mkdocs-material>=9.0.0         # Material theme for docs
mkdocs-mermaid2-plugin>=0.6.0  # Mermaid integration

# Development and testing
pytest>=7.0.0                  # Testing framework
pytest-asyncio>=0.21.0         # Async testing support
black>=22.0.0                  # Code formatting
mypy>=1.0.0                    # Type checking
coverage>=6.0                  # Code coverage

# Optional performance enhancements
uvloop>=0.17.0; sys_platform != "win32"  # Fast event loop (Unix only)
orjson>=3.8.0                  # Fast JSON parsing
aiodns>=3.0.0                  # Async DNS resolution

# Integration dependencies
requests>=2.28.0               # HTTP requests
aiohttp>=3.8.0                 # Async HTTP client
httpx>=0.24.0                  # Modern HTTP client

# Diagram and visualization
matplotlib>=3.5.0              # Plotting library
seaborn>=0.11.0                # Statistical visualization
networkx>=2.8.0                # Graph analysis

# Enterprise integration (optional)
# ldap3>=2.9.0                 # LDAP integration
# pyodbc>=4.0.0                # Database connectivity
# kubernetes>=24.0.0           # Kubernetes API client